# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Beep communities openapi",
    "description": "API documentation for the Communities service",
    "contact": {
      "name": "communities-core@beep.ovh"
    },
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.0.1"
  },
  "paths": {
...
    "/invitations/{invitation_id}": {
      "get": {
        "tags": [
          "server_invitations"
        ],
        "operationId": "get_invitation",
        "parameters": [
          {
            "name": "invitation_id",
            "in": "path",
            "description": "Invitation ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invitation retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerInvitation"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Invitation not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/invitations/{invitation_id}/accept": {
      "post": {
        "tags": [
          "server_invitations"
        ],
        "operationId": "accept_invitation",
        "parameters": [
          {
            "name": "invitation_id",
            "in": "path",
            "description": "Invitation ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invitation accepted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not the invitee or invitation expired"
          },
          "404": {
            "description": "Invitation not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/servers/{id}": {
      "get": {
        "tags": [
          "servers"
        ],
        "operationId": "get_server",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Server"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Server is private"
          },
          "404": {
            "description": "Server not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "tags": [
          "servers"
        ],
        "operationId": "update_server",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Server updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Server"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid server name"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not the server owner"
          },
          "404": {
            "description": "Server not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "servers"
        ],
        "operationId": "delete_server",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not the server owner"
          },
          "404": {
            "description": "Server not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/servers/{server_id}/invitations": {
      "post": {
        "tags": [
          "server_invitations"
        ],
        "operationId": "create_invitation",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServerInvitationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Invitation created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerInvitation"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not server owner"
          },
          "404": {
            "description": "Server not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateServerInvitationRequest": {
        "type": "object",
        "properties": {
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "invitee_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UserId"
              }
            ]
          }
        }
      },
...
      "Server": {
        "type": "object",
        "required": [
          "id",
          "name",
          "owner_id",
          "visibility",
          "created_at"
        ],
        "properties": {
          "banner_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "$ref": "#/components/schemas/ServerId"
          },
          "name": {
            "type": "string"
          },
          "owner_id": {
            "$ref": "#/components/schemas/UserId"
          },
          "picture_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "updated_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "visibility": {
            "$ref": "#/components/schemas/ServerVisibility"
          }
        }
      },
      "ServerId": {
        "type": "string",
        "format": "uuid"
      },
      "ServerInvitation": {
        "type": "object",
        "required": [
          "id",
          "server_id",
          "inviter_id",
          "status",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "id": {
            "$ref": "#/components/schemas/ServerInvitationId"
          },
          "invitee_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UserId"
              }
            ]
          },
          "inviter_id": {
            "$ref": "#/components/schemas/UserId"
          },
          "server_id": {
            "$ref": "#/components/schemas/ServerId"
          },
          "status": {
            "$ref": "#/components/schemas/ServerInvitationStatus"
          },
          "updated_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "ServerInvitationId": {
        "type": "string",
        "format": "uuid"
      },
      "ServerInvitationStatus": {
        "type": "string",
        "enum": [
          "Pending",
          "Accepted",
          "Rejected"
        ]
      },
...
      "ServerVisibility": {
        "type": "string",
        "enum": [
          "Public",
          "Private"
        ]
      },
...
      "UpdateServerRequest": {
        "type": "object",
        "properties": {
          "banner_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "picture_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "visibility": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ServerVisibility"
              }
            ]
          }
        }
      },
      "UserId": {
        "type": "string"
      },
...
    }
  }
}